Fixing issue with new Chrome and window.open
In Chrome 19+, new tabs are treated like they belong ot the "newtab" domain. When you try and launch an app from another you cannot have a handle the window (cross domain javascript limitations). To get around this, I'm launching the address of the app and then launching another window on the same domain so that we can control it. Tests have started to work again.